table_creator/requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
tqdm
python-dotenv
numpy
torch
transformers
scipy
scikit-image
opencv-python-headless
mercantile
mapbox_vector_tile
vt2geojson
pillow
requests
imagesize
torchvision
matplotlib